+++ Sources/RepoBar/Support/AppActions.swift
@@ -4,9 +4,7 @@ import AppKit
 enum AppActions {
     static func openSettings() {
         NSApp.activate(ignoringOtherApps: true)
-        if !NSApp.sendAction(Selector(("showSettingsWindow:")), to: nil, from: nil) {
-            _ = NSApp.sendAction(Selector(("showPreferencesWindow:")), to: nil, from: nil)
-        }
+        NotificationCenter.default.post(name: .repoBarOpenSettings, object: nil)
     }
 
     static func openAbout() {